# perf(prom): optimize label values query