--- a/src/query/src/promql/planner.rs
+++ b/src/query/src/promql/planner.rs
@@ -939,7 +939,7 @@ impl PromPlanner {
             Some(Offset::Neg(duration)) => -(duration.as_millis() as Millisecond),
             None => 0,
         };
-        let mut scan_filters = self.matchers_to_expr(label_matchers.clone(), table_schema)?;
+        let mut scan_filters = Self::matchers_to_expr(label_matchers.clone(), table_schema)?;
         if let Some(time_index_filter) = self.build_time_index_filter(offset_duration)? {
             scan_filters.push(time_index_filter);
         }
@@ -1135,8 +1135,7 @@ impl PromPlanner {
     }
 
     // TODO(ruihang): ignore `MetricNameLabel` (`__name__`) matcher
-    fn matchers_to_expr(
-        &self,
+    pub fn matchers_to_expr(
         label_matchers: Matchers,
         table_schema: &DFSchemaRef,
     ) -> Result<Vec<DfExpr>> {