I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)
#5525 

## What's changed and what's your intention?

This PR optimizes the label values query by using a specified logical plan, eliminating unnecessary overhead from the PromQL plan. This change improves performance by avoiding extra operations like sorting labels and dividing timeseries.

1h time range
PromQL 1.21s vs SQL 0.29s

24h time range
PromQL 4.32s vs SQL 0.75s
```
tql eval(1741183200,1741186800,300) up{job="node_exporter"};

|     0 |    0 |  PromInstantManipulateExec: range=[1741183200000..1741186800000], lookback=[300000], interval=[300000], time index=[greptime_timestamp] metrics=[output_rows: 13057, elapsed_compute: 208885377, num_series: 1057, ]
  PromSeriesNormalizeExec: offset=[0], time index=[greptime_timestamp], filter NaN: [false] metrics=[output_rows: 1400003, elapsed_compute: 203537424, num_series: 1057, ]
    PromSeriesDivideExec: tags=["instance", "job", "replica", "revision"] metrics=[output_rows: 1400003, elapsed_compute: 150884149, num_series: 1057, ]
      SortPreservingMergeExec: [instance@2 ASC NULLS LAST, job@3 ASC NULLS LAST, replica@4 ASC NULLS LAST, revision@5 ASC NULLS LAST] metrics=[output_rows: 1400003, elapsed_compute: 106749221, ]
        SortExec: expr=[instance@2 ASC NULLS LAST, job@3 ASC NULLS LAST, replica@4 ASC NULLS LAST, revision@5 ASC NULLS LAST], preserve_partitioning=[true] metrics=[output_rows: 1400003, elapsed_compute: 244091535, spill_count: 0, spilled_bytes: 0, spilled_rows: 0, ]
          MergeScanExec: peers=[5205500362752(1212, 0), ] metrics=[output_rows: 1400003, greptime_exec_read_cost: 0, finish_time: 873571638, first_consume_time: 564254392, ready_time: 9807817, ]
 |
|     1 |    0 |  SortPreservingMergeExec: [instance@2 DESC NULLS LAST, job@3 DESC NULLS LAST, replica@4 DESC NULLS LAST, revision@5 DESC NULLS LAST, greptime_timestamp@0 DESC NULLS LAST] metrics=[output_rows: 1400003, elapsed_compute: 207254345, ]
  SortExec: expr=[instance@2 DESC NULLS LAST, job@3 DESC NULLS LAST, replica@4 DESC NULLS LAST, revision@5 DESC NULLS LAST, greptime_timestamp@0 DESC NULLS LAST], preserve_partitioning=[true] metrics=[output_rows: 1400003, elapsed_compute: 841183866, spill_count: 0, spilled_bytes: 0, spilled_rows: 0, ]
    CoalesceBatchesExec: target_batch_size=8192 metrics=[output_rows: 1400003, elapsed_compute: 2044944, ]
      FilterExec: greptime_timestamp@0 >= 1741182900000 AND greptime_timestamp@0 <= 1741187100000 metrics=[output_rows: 1400003, elapsed_compute: 18662743, ]
        SeqScan: region=4402341478400(1025, 0), partition_count=25621 (0 memtable ranges, 4 file 25622 ranges) metrics=[output_rows: 1400003, mem_used: 384000160, elapsed_await: 5013476215, elapsed_poll: 4938347156, ]
      |
|  NULL | NULL | Total rows: 13057   
```

```
SELECT DISTINCT instance FROM up WHERE job = 'node_exporter'  AND greptime_timestamp BETWEEN '2025-03-05T14:00:00Z' AND '2025-03-05T15:00:00Z';

|     0 |    0 |  AggregateExec: mode=FinalPartitioned, gby=[instance@0 as instance], aggr=[] metrics=[output_rows: 1000, elapsed_compute: 239756, spill_count: 0, spilled_bytes: 0, spilled_rows: 0, peak_mem_used: 91456, ]
  CoalesceBatchesExec: target_batch_size=8192 metrics=[output_rows: 1000, elapsed_compute: 12106, ]
    RepartitionExec: partitioning=Hash([instance@0], 8), input_partitions=8 metrics=[fetch_time: 289500365, repartition_time: 60229, send_time: 11761, ]
      AggregateExec: mode=Partial, gby=[instance@0 as instance], aggr=[] metrics=[output_rows: 1000, elapsed_compute: 11029704, spill_count: 0, spilled_bytes: 0, spilled_rows: 0, skipped_aggregation_rows: 0, peak_mem_used: 384540, ]
        MergeScanExec: peers=[5205500362752(1212, 0), ] metrics=[output_rows: 1200002, greptime_exec_read_cost: 0, finish_time: 289292540, first_consume_time: 207959152, ready_time: 15636605, ]
 |
|     1 |    0 |  CoalesceBatchesExec: target_batch_size=8192 metrics=[output_rows: 1200002, elapsed_compute: 478396, ]
  FilterExec: greptime_timestamp@0 >= 1741183200000 AND greptime_timestamp@0 <= 1741186800000, projection=[instance@1] metrics=[output_rows: 1200002, elapsed_compute: 8113743, ]
    SeqScan: region=4402341478400(1025, 0), partition_count=14162 (0 memtable ranges, 2 file 14162 ranges) metrics=[output_rows: 1200002, mem_used: 52837264, elapsed_await: 2511100886, elapsed_poll: 2755889754, ]
                                                                                                                                                                                                                                                                                                                                                                                                        |
|  NULL | NULL | Total rows: 1000
```



## PR Checklist
Please convert it to a draft if some of the following conditions are not met.

- [ ] I have written the necessary rustdoc comments.
- [ ] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.
- [ ] API changes are backward compatible.
- [ ] Schema or data changes are backward compatible.